Bug 1816879 - Move credit cards autofill related UI tests to a new class
