+++ fenix/app/src/androidTest/java/org/mozilla/fenix/ui/CreditCardAutofillTest.kt
@@ -1,0 +1,97 @@
+package org.mozilla.fenix.ui
+
+import okhttp3.mockwebserver.MockWebServer
+import org.junit.After
+import org.junit.Before
+import org.junit.Rule
+import org.junit.Test
+import org.mozilla.fenix.customannotations.SmokeTest
+import org.mozilla.fenix.helpers.AndroidAssetDispatcher
+import org.mozilla.fenix.helpers.HomeActivityIntentTestRule
+import org.mozilla.fenix.helpers.TestAssetHelper
+import org.mozilla.fenix.ui.robots.homeScreen
+import org.mozilla.fenix.ui.robots.navigationToolbar
+import java.time.LocalDate
+
+class CreditCardAutofillTest {
+    private lateinit var mockWebServer: MockWebServer
+
+    object CreditCard {
+        const val MOCK_CREDIT_CARD_NUMBER = "5555555555554444"
+        const val MOCK_LAST_CARD_DIGITS = "4444"
+        const val MOCK_NAME_ON_CARD = "Mastercard"
+        const val MOCK_EXPIRATION_MONTH = "February"
+        val MOCK_EXPIRATION_YEAR = (LocalDate.now().year + 1).toString()
+    }
+
+    @get:Rule
+    val activityIntentTestRule = HomeActivityIntentTestRule.withDefaultSettingsOverrides()
+
+    @Before
+    fun setUp() {
+        mockWebServer = MockWebServer().apply {
+            dispatcher = AndroidAssetDispatcher()
+            start()
+        }
+    }
+
+    @After
+    fun tearDown() {
+        mockWebServer.shutdown()
+    }
+
+    @SmokeTest
+    @Test
+    fun verifyCreditCardAutofillTest() {
+        val creditCardFormPage = TestAssetHelper.getCreditCardFormAsset(mockWebServer)
+
+        homeScreen {
+        }.openThreeDotMenu {
+        }.openSettings {
+        }.openAutofillSubMenu {
+            clickAddCreditCardButton()
+            fillAndSaveCreditCard(
+                CreditCard.MOCK_CREDIT_CARD_NUMBER,
+                CreditCard.MOCK_NAME_ON_CARD,
+                CreditCard.MOCK_EXPIRATION_MONTH,
+                CreditCard.MOCK_EXPIRATION_YEAR,
+            )
+            // Opening Manage saved cards to dismiss here the Secure your credit prompt
+            clickManageSavedCardsButton()
+            clickSecuredCreditCardsLaterButton()
+        }.goBackToAutofillSettings {
+        }.goBack {
+        }.goBack {
+        }
+        navigationToolbar {
+        }.enterURLAndEnterToBrowser(creditCardFormPage.url) {
+            clickCardNumberTextBox()
+            clickSelectCreditCardButton()
+            clickCreditCardSuggestion(CreditCard.MOCK_LAST_CARD_DIGITS)
+            verifyAutofilledCreditCard(CreditCard.MOCK_CREDIT_CARD_NUMBER)
+        }
+    }
+
+    @SmokeTest
+    @Test
+    fun deleteSavedCreditCardTest() {
+        homeScreen {
+        }.openThreeDotMenu {
+        }.openSettings {
+        }.openAutofillSubMenu {
+            clickAddCreditCardButton()
+            fillAndSaveCreditCard(
+                CreditCard.MOCK_CREDIT_CARD_NUMBER,
+                CreditCard.MOCK_NAME_ON_CARD,
+                CreditCard.MOCK_EXPIRATION_MONTH,
+                CreditCard.MOCK_EXPIRATION_YEAR,
+            )
+            clickManageSavedCardsButton()
+            clickSecuredCreditCardsLaterButton()
+            clickSavedCreditCard()
+            clickDeleteCreditCardButton()
+            clickConfirmDeleteCreditCardButton()
+            verifyAddCreditCardsButton()
+        }
+    }
+}
